Fix build.
Specifically:

  external/webrtc/src/modules/audio_coding/codecs/isac/fix/source/../test/kenny.c:366: error: undefined reference to 'atof'

The real problem here is that this code is compiling against the platform
bionic headers but linking against the NDK C library. That's a terrible
idea, but I can't understand the makefile well enough to stop that.

This fixes the build by using a function that's common to both C libraries.

Change-Id: I3d5747014a45269520a4c3585be42f2a78293d86