Remove __DATE__ and __TIME__ from tracing
Including __DATE__ and __TIME__ in the binary means that every build of
this library is different, even if the sources are identical. It also
means that every OTA needs to patch this library. Remove __DATE__ and
__TIME__ from the trace build info, just record whether it's a debug or
release build.

Bug: 24204119
Change-Id: I0dfc1479f096197007db1920b89aeb61ea537168
(cherry picked from commit e3422c11696e715e66b1669230410f098bf4c038)